clean up adapter
